feat(escpos): enable ability to disable beeper, and support legacy raster print
Some "ESCPOS Compatible" printers do not have support for beepers, and instead
print the escape sequence at the top of the docket. This commit enables this
setting to be disabled.

Older model ESCPOS printers do not suppor the newer raster print commands.
Add support for the legacy/obsolete modes to support older printers that
are still in circulation.